Refactor SPICE geometry into separate module
