+++ mmx_phobos_codes/Phobos_HRSC.py
@@ -7,10 +7,7 @@ import os
 sys.path.append(os.getcwd())
 import cortopy as corto
 import numpy as np
-import spiceypy as spice
-from scipy.spatial.transform import Rotation as R
-import time
-import requests
+from mmx_phobos_codes import phobos_geometry
 
 ## Clean all existing/Default objects in the scene 
 corto.Utils.clean_scene()
@@ -33,76 +30,10 @@ State.add_path('uv_data_path_2',os.path.join(State.path["input_path"],'body','uv
 State.add_path('albedo_path_3',os.path.join(State.path["input_path"],'body','albedo','Deimos grayscale.jpg'))
 State.add_path('uv_data_path_3',os.path.join(State.path["input_path"],'body','uv data','g_deimos_162m_spc_0000n00000_v001.json'))
 
-### (1.1) LOAD SPICE KERNELS AND COMPUTE GEOMETRY ###
-UTC_TIME = "2018-08-02T08:48:03.686"
-
-KERNEL_LIST = {
-    "generic_kernels/lsk/naif0012.tls",
-    "generic_kernels/pck/pck00010.tpc",
-    "MEX/kernels/sclk/former_versions/MEX_250417_STEP.TSC",
-    "MEX/kernels/spk/MAR097_030101_300101_V0001.BSP",
-    "generic_kernels/spk/planets/a_old_versions/de405.bsp",
-    "MEX/kernels/spk/ORMM_T19_180801000000_01460.BSP",
-    "MEX/kernels/ck/ATNM_MEASURED_180101_181231_V01.BC",
-    "MEX/kernels/fk/MEX_V16.TF",
-}
-
-MIRRORS = [
-    "https://naif.jpl.nasa.gov/pub/naif",
-    "http://naif.jpl.nasa.gov/pub/naif",
-    "https://spiftp.esac.esa.int/data/SPICE",
-]
-
-def dl(rel, mirrors, ret=3):
-    fn = os.path.basename(rel)
-    if os.path.isfile(fn) and os.path.getsize(fn):
-        return
-    for m in mirrors:
-        url = f"{m}/{rel}"
-        for n in range(1, ret + 1):
-            try:
-                with requests.get(url, timeout=(10, 300), stream=True) as r:
-                    r.raise_for_status()
-                    with open(fn, "wb") as f:
-                        for c in r.iter_content(1 << 20):
-                            f.write(c)
-                return
-            except (requests.Timeout, requests.ConnectionError):
-                time.sleep(2)
-    raise RuntimeError(f"{fn} could not be downloaded")
-
-for k in KERNEL_LIST:
-    dl(k, MIRRORS)
-for fn in map(os.path.basename, KERNEL_LIST):
-    spice.furnsh(fn)
-
-et = spice.str2et(UTC_TIME)
-v_pho_mars = spice.spkezr("PHOBOS", et, "J2000", "NONE", "MARS")[0][:3]
-v_dei_mars = spice.spkezr("DEIMOS", et, "J2000", "NONE", "MARS")[0][:3]
-v_mex_mars = spice.spkezr("-41", et, "J2000", "NONE", "MARS")[0][:3]
-v_sun_mars = spice.spkezr("SUN", et, "J2000", "NONE", "MARS")[0][:3]
-
-R_i_mars = spice.pxform("J2000", "IAU_MARS", et)
-R_i_pho = spice.pxform("J2000", "IAU_PHOBOS", et)
-R_i_dei = spice.pxform("J2000", "IAU_DEIMOS", et)
-R_i_sc = spice.pxform("J2000", "MEX_SPACECRAFT", et)
-
-v_sc_pho = R_i_pho @ (v_mex_mars - v_pho_mars)
-v_mar_pho = R_i_pho @ (-v_pho_mars)
-v_dei_pho = R_i_pho @ (v_dei_mars - v_pho_mars)
-v_sun_pho = R_i_pho @ (v_sun_mars - v_pho_mars)
-
-R_pho_sc = R_i_sc @ R_i_pho.T
-R_pho_mars = R_i_mars @ R_i_pho.T
-R_pho_dei = R_i_dei @ R_i_pho.T
-
-def rot_to_quat(Rm):
-    q = R.from_matrix(Rm).as_quat()  # returns [x,y,z,w]
-    return np.array([q[3], q[0], q[1], q[2]])
-
-quat_sc = rot_to_quat(R_pho_sc)
-quat_mars = rot_to_quat(R_pho_mars)
-quat_dei = rot_to_quat(R_pho_dei)
+### (1.1) LOAD SPICE GEOMETRY ###
+geometry = phobos_geometry.compute_geometry()
+positions = geometry["positions"]
+orientations = geometry["orientations"]
 
 ### (2) SETUP THE SCENE ###
 # Setup bodies
@@ -182,15 +113,15 @@ n_img = 1
 for idx in range(n_img):
     body_1.set_position(np.array([0, 0, 0]))
     body_1.set_orientation(np.array([1, 0, 0, 0]))
-    body_2.set_position(v_mar_pho)
-    body_2.set_orientation(quat_mars)
-    body_3.set_position(v_dei_pho)
-    body_3.set_orientation(quat_dei)
-    cam.set_position(v_sc_pho)
-    cam.set_orientation(quat_sc)
-    cam_hrsc.set_position(v_sc_pho)
-    cam_hrsc.set_orientation(quat_sc)
-    sun.set_position(v_sun_pho)
+    body_2.set_position(positions["mars"])
+    body_2.set_orientation(orientations["mars"])
+    body_3.set_position(positions["deimos"])
+    body_3.set_orientation(orientations["deimos"])
+    cam.set_position(positions["sc"])
+    cam.set_orientation(orientations["sc"])
+    cam_hrsc.set_position(positions["sc"])
+    cam_hrsc.set_orientation(orientations["sc"])
+    sun.set_position(positions["sun"])
     ENV.RenderOne(cam, State, index=idx, depth_flag=True)
     ENV.RenderOne(cam_hrsc, State, index=idx, depth_flag=True)
 
@@ -198,5 +129,4 @@ for idx in range(n_img):
 corto.Utils.save_blend(State)
 
 # Unload kernels
-for fn in map(os.path.basename, KERNEL_LIST):
-    spice.unload(fn)
+phobos_geometry.unload_kernels()
